+++ cmd/magebox/root.go
@@ -10,7 +10,7 @@ import (
 	"github.com/qoliber/magebox/internal/verbose"
 )
 
-var version = "0.17.2"
+var version = "0.17.3"
 
 // verbosity is the count of -v flags
 var verbosity int